==> scripts/05.2_Assemble_explore_modeling_results.R <==


###### Script 5_B: Assemble and explore modeling results from the cluster #######

# To gather the several summary tables generated by the parallelized version of Script 05.1a and assemble them

# To explore the results


# Remove environment
rm(list = ls())

# Load libraries
library(biomod2)
library(tidyverse)

## Select Env data resolution 
# res <- "5"
res <- "15"

##### 1/ Assemble the results from the cluster: List.models #####


rds_files_path <- list.files(path = paste0("./models/",res,"/0_List.models_outputs"), pattern = "(\\.rds)$", ignore.case = T, full.names = T)
rds_files_names <- list.files(path = paste0("./models/",res,"/0_List.models_outputs"), pattern = "(\\.rds)$", ignore.case = T, full.names = F)

# Extract indices for computed models in each .rds
indices <- str_extract_all(rds_files_names, pattern = "[0123456789]{1,}")

list.models.to.build <- vector(mode = "list", length = length(rds_files_path))
for (i in seq_along(rds_files_path)) {
  list.models.to.build[[i]] <- readRDS(file = rds_files_path[i])[indices[[i]][1]:indices[[i]][2], ] # Extract only rows for OMU modeled in this portion
}

list.models <- list.models.to.build %>% 
  reduce(.f = rbind) %>% 
  arrange(ID_unit)

# Explore results
hist(list.models$Models_success)
hist(list.models$N.obs_15m_used[list.models$Models_success < 27])

hist(list.models$Models_success[list.models$N.obs_15m_used < 9])


# ### Find missing models
# 
# which((list.models$initial_model_type != "rasterized") & is.na(list.models$Start_Time))


# ### Fix error with quick evaluation of TSS ####
# 
# model.runs <- NA
# for (i in 1:nrow(list.models)) {
#   
#   # i <- 1
#   unit <- as.character(list.models$Tag.model[i])  
#   
#   initial_model_type <- list.models$initial_model_type[i]
#   
#   if (initial_model_type != "rasterized") {
#     
#     # Load model.runs
#     model.runs <- readRDS(model.runs, file = paste0("./models/",res,"/",unit,"/model.runs.rds"))
#     
#     # Get evaluations from biomod2
#     eval.runs <- subset(x=reshape2::melt(get_evaluations(model.runs, as.data.frame = F)), subset = (Var2 == "Testing.data"))[, c(3:6)]
#     
#     # Store info on Biomod2 Evaluation with TSS: Unparallelized version
#     list.models$Mean_TSS_full[list.models$Tag.model == unit] <- round(mean(eval.runs$value, na.rm = T),3) # Mean value for TSS among all converged models
#     list.models$Models_select_TSS_0.5[list.models$Tag.model == unit] <- sum(eval.runs$value >= 0.5, na.rm = T) # Nb of model with TSS ≥ 0.5
#     list.models$Mean_TSS_select_TSS_0.5[list.models$Tag.model == unit] <- round(mean(eval.runs$value[eval.runs$value >= 0.5], na.rm = T),3) # Mean value for TSS among all models with TSS ≥ 0.5
#     
#   }
#   
#   
# }

# save(list.models, file = "./input_data/list.models.RData", version = "2") ####


##### 2/ Explore model projection outputs #####


# Choice of unit
load(file = paste0("./input_data/list.models.RData"))

i <- 1
unit <- list.models$Tag.model[i]

# Choice of resolution
res <- "15"


# Loading occurrence data
load(paste0("./input_data/Species_data/",res,"/Spatial_Points_Objects/occurrences_", unit,".RData")) #

# Loading projection stack
proj_stack <- readRDS(file = paste0("./models/",res,"/",unit,"/proj_Current/proj_Current_",unit,".rds"))

# Plot all
plot(proj_stack)
points(unit.points, pch = 16)

# Plot a subset

# By PA run
subset <- "_PA1_"
# By algo
subset <- "GBM"
subset <- "ANN"
subset <- "RF"
# By CV_run (does not really make sense since they are different among PA set, but whatever)
subset <- "_RUN1_"

plot(proj_stack[[grep(pattern = subset, x = names(proj_stack))]])
points(unit.points, pch = 16)

# Loading Clamping mask
Clamping_mask <- readRDS(file = paste0("./models/",res,"/",unit,"/proj_Current/proj_Current_",unit,"_ClampingMask.rds"))

plot(Clamping_mask)
points(unit.points, pch = 16)
